# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "3X-UI Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a 3X-UI panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "3x-ui",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/panel/setting/update": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Persist every setting at once. The body mirrors the shape returned by /all. Invalid values (bad ports, missing cert pairs, etc.) are rejected before write.",
        "operationId": "post_panel_setting_update",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/setting/updateUser": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Change the panel admin username and password. Requires the current credentials for verification. The session is refreshed with the new values on success.",
        "operationId": "post_panel_setting_updateUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oldUsername": {
                    "type": "string",
                    "description": "Current admin username."
                  },
                  "oldPassword": {
                    "type": "string",
                    "description": "Current admin password."
                  },
                  "newUsername": {
                    "type": "string",
                    "description": "Desired new username."
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "Desired new password."
                  }
                },
                "required": [
                  "oldUsername",
                  "oldPassword",
                  "newUsername",
                  "newPassword"
                ]
              },
              "example": {
                "oldUsername": "admin",
                "oldPassword": "admin",
                "newUsername": "newadmin",
                "newPassword": "newpass"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/setting/apiTokens/create": {
      "post": {
        "tags": [
          "API Tokens"
        ],
        "summary": "Mint a new API token. Name must be unique and 1-64 characters; the token string is server-generated.",
        "operationId": "post_panel_setting_apiTokens_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Human-readable label, e.g. \"central-panel-a\"."
                  }
                },
                "required": [
                  "name"
                ]
              },
              "example": {
                "name": "central-panel-a"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "id": 2,
                    "name": "central-panel-a",
                    "token": "new-token-string",
                    "enabled": true,
                    "createdAt": 1736000000
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "success": false,
                  "msg": "a token with that name already exists"
                }
              }
            }
          }
        }
      }
    },
    "/panel/setting/apiTokens/delete/{id}": {
      "post": {
        "tags": [
          "API Tokens"
        ],
        "summary": "Permanently delete a token. Any caller using it stops authenticating immediately.",
        "operationId": "post_panel_setting_apiTokens_delete_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Token row ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true
                }
              }
            }
          }
        }
      }
    },
    "/panel/setting/apiTokens/setEnabled/{id}": {
      "post": {
        "tags": [
          "API Tokens"
        ],
        "summary": "Toggle a token enabled/disabled without deleting it. Disabled tokens are rejected by checkAPIAuth on the next request.",
        "operationId": "post_panel_setting_apiTokens_setEnabled_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Token row ID.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean",
                    "description": "New enabled state."
                  }
                },
                "required": [
                  "enabled"
                ]
              },
              "example": {
                "enabled": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true
                }
              }
            }
          }
        }
      }
    },
...
  }
}